requirements.txt
flask
flask-sqlalchemy
requests
Pillow
tensorflow
tf-keras
transformers
python-dotenv
werkzeug
gunicorn